% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp.R
\name{clean_comarcas_names}
\alias{clean_comarcas_names}
\title{Cleans comarca names}
\usage{
clean_comarcas_names(muni_com_comarca_raw)
}
\arguments{
\item{muni_com_comarca_raw}{object returned from \code{get_comarca_text} fun}
}
\value{
data frame
}
\description{
Cleans comarca names from \code{get_comarca_text} function.
}
