% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_sf.R
\docType{data}
\name{d_sf}
\alias{d_sf}
\title{Shapefile for jurimetric data}
\format{A data frame with 4 rows and 2 variables:
\describe{
  \item{nivel}{level of data (municipality, comarca, circunscricao, region)}
  \item{sf}{simple feature complex data}
  ...
}}
\source{
\url{http://www.tjsp.jus.br/ListaTelefonica}
}
\usage{
d_sf
}
\description{
A dataset containing shapefiles for municipalities, comarcas,
circunscricoes and regions
}
\keyword{datasets}
