% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{d_sf_tjrs}
\alias{d_sf_tjrs}
\title{Shapefile TJRS}
\format{
A data frame with 2 rows and 2 variables:
\describe{
  \item{nivel}{nivel da regiao (municipio, comarca)}
  \item{sf}{shape, num objeto do tipo simple feature}
}
}
\source{
\url{http://www.tjrs.jus.br/institu/comarcas/}
}
\usage{
d_sf_tjrs
}
\description{
Base contendo os shapefiles de municipios e comarcas do TJRS.
}
\keyword{datasets}
