% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{d_sf_tjsp}
\alias{d_sf_tjsp}
\title{Shapefile TJSP}
\format{
A data frame with 4 rows and 2 variables:
\describe{
  \item{nivel}{nivel da regiao (municipio, comarca, circunscricao, regiao)}
  \item{sf}{shape, num objeto do tipo simple feature}
}
}
\source{
\url{http://www.tjsp.jus.br/ListaTelefonica}
}
\usage{
d_sf_tjsp
}
\description{
Base contendo os shapefiles de municipios, comarcas,
regioes e circunscricoes do TJSP.
}
\keyword{datasets}
