% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_comarcas.R
\name{get_all_muni}
\alias{get_all_muni}
\title{Downloads all municipalities}
\usage{
get_all_muni()
}
\value{
\code{data.frame} with t
}
\description{
Downloads all municipalities from Lista Telefonica data
}
