% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp.R
\name{get_comarca_text}
\alias{get_comarca_text}
\title{Gets comarca text}
\usage{
get_comarca_text(cod_municipio)
}
\arguments{
\item{cod_municipio}{code}
}
\value{
returns a phrase if it is a municipality and "comarca" if it is a comarca
}
\description{
Gets comarca text from municipality code to tell whether the municipality is
a comarca or just a part of the comarca.
}
