% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp.R
\name{get_muni_codes}
\alias{get_muni_codes}
\title{Downloads the codes for all the municipalities}
\usage{
get_muni_codes()
}
\value{
\code{character} vector with all 645 cleaned municipalities
}
\description{
Downloads all municipalities codes from Lista Telefonica data - TJSP
}
