% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_sf.R
\docType{data}
\name{d_sf_tjrs}
\alias{d_sf_tjrs}
\title{Shapefile for jurimetric data}
\format{A data frame with 4 rows and 2 variables:
\describe{
  \item{nivel}{level of data (municipality, comarca)}
  \item{sf}{simple feature complex data}
  ...
}}
\source{
\url{http://www.tjrs.jus.br/institu/comarcas/}
}
\usage{
d_sf_tjrs
}
\description{
A dataset containing shapefiles for municipalities and comarcas,
}
\keyword{datasets}
