% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp.R
\name{get_all_muni}
\alias{get_all_muni}
\title{Downloads all municipalities}
\usage{
get_all_muni()
}
\value{
\code{character} vector with all 645 cleaned municipalities
}
\description{
Downloads all municipalities from Lista Telefonica data - TJSP
}
